// src/ui/services/lending/types.ts
export type AuthorizationsStatus =
  | 'AUTHORIZATION_STATUS_PENDING'
  | 'AUTHORIZATION_STATUS_AUTHORIZED'
  | 'AUTHORIZATION_STATUS_REJECTED';
// ...
export enum LoanStatusEnum {
  Unspecified,
  Requested,
  Cancelled,
  Authorized,
  Rejected,
  Open,
  Repaid,
  Defaulted,
  Liquidated,
  Closed
}

export enum LiquidationStatusEnum {
  LIQUIDATION_STATUS_UNSPECIFIED,
  LIQUIDATION_STATUS_LIQUIDATING,
  LIQUIDATION_STATUS_LIQUIDATED,
  LIQUIDATION_STATUS_SETTLING,
  LIQUIDATION_STATUS_SETTLED
}

export type LoanStatus = keyof typeof LoanStatusEnum;
export type LiquidationStatus = keyof typeof LiquidationStatusEnum;

export interface Loan {
  borrow_amount: {
    amount: string;
    denom: string;
  };
  borrow_apr: number;
  borrower: string;
  borrowerPubKey: string;
  collateral_amount: string;
  create_at: number;
  dcm: string;
  authorizations: {
    deposit_txs: string[];
    id: string;
    status: AuthorizationsStatus;
  }[];
  disburse_at: number;
  final_timeout: string;
  interest: string;
  dlc_event_id: string;
  liquidation_id: string;
  liquidation_price: string;
  maturity: string;
  maturity_time: string;
  min_maturity: string;
  origination_fee: string;
  pool_id: string;
  protocol_fee: string;
  referrer: {
    address: string;
    name: string;
    referral_code: string;
    referral_fee_factor: string;
  };
  request_fee: {
    amount: string;
    denom: string;
  };
  status: LoanStatus;
  vault_address: string;
}